Services/emergency/requirements.txt
flask
pyserial
paho-mqtt
requests
